@@ -1,16 +1,18 @@
 name: authentication_repository
 description: Dart package which manages the authentication domain.
-version: 1.0.0
+version: 2.0.0+1
+homepage: https://github.com/tedcrimson/authentication_repository
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.7.0 <3.0.0"
+  flutter: ">=1.17.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  meta: ^1.1.8
-  formz: ^0.3.1
-  bloc: ^6.0.3
-  flutter_bloc: ^6.0.6
-  equatable: ^1.2.5
+  meta: ^1.3.0
+  formz: ^0.3.2
+  bloc: ^7.0.0
+  flutter_bloc: ^7.0.0
+  equatable: ^2.0.0
