@@ -1,6 +1,7 @@
 part of 'full_name_sign_up_cubit.dart';
 
-class FullNameSignUpState<T extends EmailForm, P extends PasswordForm> extends AuthState {
+class FullNameSignUpState<T extends EmailForm, P extends PasswordForm>
+    extends AuthState {
   factory FullNameSignUpState({
     @required String firstNameKey,
     @required String lastNameKey,
@@ -46,8 +47,10 @@ class FullNameSignUpState<T extends EmailForm, P extends PasswordForm> extends A
   }) : super({
           firstNameKey: FormEntity(firstName, StringFieldForm.constructor),
           lastNameKey: FormEntity(lastName, StringFieldForm.constructor),
-          emailKey: FormEntity(email, emailDirtyFunction ?? EmailForm.constructor),
-          passwordKey: FormEntity(password, passwordDirtyFunction ?? PasswordForm.constructor),
+          emailKey:
+              FormEntity(email, emailDirtyFunction ?? EmailForm.constructor),
+          passwordKey: FormEntity(
+              password, passwordDirtyFunction ?? PasswordForm.constructor),
         }, status, autoValidate);
 
   final T email;
