@@ -20,7 +20,8 @@ class PasswordForm extends StringFieldForm {
   const PasswordForm.dirty({String value, bool requiredField})
       : super.dirty(value: value, requiredField: requiredField);
 
-  const PasswordForm.pure({bool requiredField = false}) : super.pure(requiredField);
+  const PasswordForm.pure({bool requiredField = false})
+      : super.pure(requiredField);
   static constructor(String value, bool requiredField) =>
       PasswordForm.dirty(value: value, requiredField: requiredField);
 
